refactor(renderers): share one orthogonal connection router
The candidate families, the endpoint-side contract, and the obstacle-clearing
tests lived inside the architecture renderer, so a second type that needs the
same routing could only copy them. Move them into
renderers/shared/orthogonal-router.mjs and let createOrthogonalRouter own port
spreading, side selection, and the route cache.

The extraction adds two opt-in hooks that are unused by the existing callers:
preferredCandidates runs before the shared families (a type can own the channel
choice for relationships that would otherwise share one corridor) and
extraCandidates runs after them, before the best-effort fallback. strictClearance
lets a type that can route around obstacles take the direct line only when that
line is genuinely clear; the default keeps the historical behaviour where the
Clean Flow gate names the obstacle.

No behaviour change for existing types: architecture's web-app example renders
byte-identical to the checked-in golden.
